Tree: refactor tree
